finished backend

+++ fitback/src/main/java/com/smeekens/fitback/fitback/fitback/security/services/FeedbackService.java
@@ -40,7 +40,11 @@ public class FeedbackService {
     }
 
     public void deleteFeedback(Long id) {
-        feedbackRepository.deleteById(id);
+        if (feedbackRepository.existsById(id)) {
+            feedbackRepository.deleteById(id);
+        } else {
+            throw new RecordNotFoundException();
+        }
 
     }
